Create getOnlyChanges.sh
This gives a zip file containing all the changes between two commits so that you can upload (via FTP for example) to a remote server. This is especially useful if you cannot use rsync or git to deploy.
After zipping up the changed files, the script echos a raw list of files that need to be deleted.